fix(sound): WaveMixer GetLatencySamples 负索引走错分支越界读
idx 为 int、_bufferCount 为 tjs_uint,比较时 idx 被提升为无符号,
负 idx 变巨大正数走 idx -= _bufferCount 分支,_bufferSize[负] 越界读
(Debug Asan 实锤 heap-buffer-overflow)。给 _bufferCount 加 (int) cast。

tTVPSoundBufferAL(OpenAL) 是 Web 专用后端,libkrkr2.so 用
Oboe/OpenSL 软混音,无对应物(平台边界)。

+++ cpp/core/sound/win32/WaveMixer.cpp
@@ -694,7 +694,10 @@ public:
         tjs_int total = -offset;
         for(int i = 0; i < remainBuffers; ++i) {
             int idx = _bufferIdx + 1 - remainBuffers + i;
-            if(idx >= _bufferCount)
+            // _bufferCount is unsigned; without the cast a negative idx is
+            // promoted to a huge unsigned value, takes the wrong branch and
+            // indexes _bufferSize out of bounds (heap-buffer-overflow).
+            if(idx >= (int)_bufferCount)
                 idx -= _bufferCount;
             else if(idx < 0)
                 idx += _bufferCount;
